refactor(loonfs): split the runtime facade into focused modules
Pure code movement plus one mechanical pattern change; no functional
change. loonfs/src/lib.rs had grown to 2,400 lines mixing the public
surface with three cache implementations. It is now a 71-line crate root
(docs, re-exports, error type) over: fs (the Fs handle, builder, and the
entire public operation surface in one place), cache (verified-basis,
commit-engine, and control-object caches plus the cache-management half
of Fs), uploads (durable-content proofs), options, config, and time (the
runtime's only clock touchpoints).

The 18 repeated `.lock().expect("... lock poisoned")` call sites collapse
into three documented lock accessors on FsInner, the single place the
poisoning policy now lives. Every public path is preserved by crate-root
re-exports.
